Mitigate race conditions
At least I hope that's what this does. Even if this works as I think it
should, there's still might be a possibility that
`guild_member_addition` is not allowed to lock `member_locks` before
`guild_member_update` is called which then gets to the lock first. Even
if that is possible, it almost certainly will never happen.

If it does happen, the user who just rejoined doesn't get their roles
back and the role they got immediately upon joining overwrites their
previous roles on the database — not exactly a desirably scenario.